Keep the original physical bytes for literals
When reformatting a source, we don't want to replace the escape
sequences with the equivalent UTF8 encoded runes, we must respect
what the user originally typed in his or her editor.

scanner: Functions, which interpret escape sequences are simplified
only to scan over.
parser: ASTs for literals keep the original uninterpreted source
bytes as a slice.
scanfilt: Escapong output is no longer necessary.