golang/parser: Create assignment statements for initialization of package-level variables
For each package-level variable, which is not zero-initialized, create
an assignment statement. All such assignment statements are kept in the
package struct and serve as starting points for emitting package
initialization code and determining the correct order for variables
initialization. Each variable contains a pointer to its initialization
statement: this will be used for inferring variable type and as an edge
in the initialization dependency graph.