Allow a typename to be used before a type declaration is read from the PDB
The declarations are written to the PDB in alphabetical order. If the an
earlier declaration refers to a `TypeDecl` later in the order, reading
fails with a `BadFile` error.

Fix this by creating an "empty" `TypeDecl` struct when we cannot find a
type name in the current package. Fill this struct at the time we read
the corresponding type declaration.