Expressions refer to declarations via an OperandName
The exrta indirection is instriduced in order to be able to track the
location of the reference to a particular declaration, e.g. in
diagnostics it is needed to refer to the place where a variable is used;
in the old design only the location where a variable is declared was
available.