Redesign the type inference algorithm
Avoid spurious type inference dependencies, caused the by the
depth-first traversal of the AST, by lazily inferring the type of
subexpressions, whenever the type of the whole expression does not
depend on the type of operands. Currently these are:
  * compouns literals: have the literal type
  * call expressions: type depends on the type of the called expression,
    regardless of argument types
  * conversions: have the conversion destination type
  * type assertions: have the asserted type
  * index expressions: have the element type of the indexed expression
  * slice expressions: have the element type of the sliced expression
  * comparison expressions: always yield an untyped boolean value

The type inference pass over the AST is followed by a second one, which
performs semantics check and constant expression evaluation.