Allow use of type assertion, map index and receive expression in a single-value context
These expressions have a type in the form of the tuple
`<T,untyped-boolean>`. This tuple is considered "non-strict" in the
sense that the boolean value may be ignored whne the expression is used
in a single-value context. Contrast with the return type of a
multi-value function: it's again a tuple, but a "strict" one.